Extract S3Storage::Null into its own file
