+++ lib/s3_storage.rb
@@ -1,19 +1,7 @@
+require 's3_storage/null'
+
 class S3Storage
   NotConfiguredError = Class.new(StandardError)
-
-  NOT_CONFIGURED_ERROR_MESSAGE = 'AWS S3 bucket not correctly configured'.freeze
-
-  class Null
-    def save(_asset, _options = {}); end
-
-    def load(_asset)
-      raise NotConfiguredError.new(NOT_CONFIGURED_ERROR_MESSAGE)
-    end
-
-    def presigned_url_for(_asset, _http_method: 'GET')
-      raise NotConfiguredError.new(NOT_CONFIGURED_ERROR_MESSAGE)
-    end
-  end
 
   def self.build(bucket_name)
     bucket_name.present? ? new(bucket_name) : Null.new
